:recycle: Manage CardData with ContextAPI
